need to do the booking widget

+++ db/migrate/20230814084031_create_ships.rb
@@ -5,7 +5,7 @@ class CreateShips < ActiveRecord::Migration[7.0]
       t.string :location
       t.string :description
       t.integer :capacity
-      t.float :price_per_day
+      t.integer :price_per_day
       t.string :ship_class
       t.string :ship_origin
       t.references :user, null: false, foreign_key: true
